+++ services/support/mgmt/2020-04-01/support/CHANGELOG.md
@@ -1,2 +1,17 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Funcs
+
+1. BaseClient.TicketResourceIDLookUp(context.Context, TicketResourceIDLookupRequest) (TicketResourceIDLookUpResponse, error)
+1. BaseClient.TicketResourceIDLookUpPreparer(context.Context, TicketResourceIDLookupRequest) (*http.Request, error)
+1. BaseClient.TicketResourceIDLookUpResponder(*http.Response) (TicketResourceIDLookUpResponse, error)
+1. BaseClient.TicketResourceIDLookUpSender(*http.Request) (*http.Response, error)
+
+### Struct Changes
+
+#### New Structs
+
+1. TicketResourceIDLookUpResponse
+1. TicketResourceIDLookupRequest